# docs/swagger.yaml
definitions:
...
  user.ChatConfig:
    properties:
      frequencyPenalty:
        default: 0
        description: 话题新鲜度,-2.0-2.0,默认为0
        type: number
      model:
        description: 模型
        type: string
      temperature:
        default: 1
        description: 随机性0-2,默认为1
        type: number
    type: object
  user.CustomConfig:
    properties:
      chatConfig:
        allOf:
        - $ref: '#/definitions/user.ChatConfig'
        description: 聊天配置
      imageConfig:
        allOf:
        - $ref: '#/definitions/user.ImageConfig'
        description: 图片配置
    type: object
...
  user.ImageConfig:
    properties:
      model:
        description: 模型
        type: string
      "n":
        default: 1
        description: 返回几张图，默认1张
        type: integer
      size:
        description: 图片大小,256x256/512x512/1024x1024
        type: string
    type: object
...
  user.ProfileRequest:
    properties:
      avatar:
        description: 头像
        type: string
      customConfig:
        allOf:
        - $ref: '#/definitions/user.CustomConfig'
        description: 自定义配置
      nickName:
        description: 昵称
        type: string
    type: object
  user.ProfileResponse:
    properties:
      avatar:
        description: 头像
        type: string
      email:
        description: 邮箱
        type: string
      integral:
        description: 用户当前积分
        type: integer
      inviteCode:
        description: 邀请码
        type: string
      nickName:
        description: 昵称
        type: string
      vipUser:
        description: 是否是vip用户
        type: boolean
    type: object
  user.RegisterRequest:
    properties:
      code:
        description: 验证码
        type: string
      email:
        description: 邮箱
        type: string
      inviteCode:
        description: 邀请码
        type: string
      password:
        description: 密码
        type: string
      type:
        default: email
        description: 可选字段，默认为email
        type: string
    required:
    - code
    - email
    - password
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/user/profile:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.ProfileResponse'
      summary: 获取用户信息
      tags:
      - 用户相关接口
    post:
      parameters:
      - description: 用户信息和自定义配置
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/user.ProfileRequest'
      responses: {}
      summary: 修改用户信息
      tags:
      - 用户相关接口
  /api/v1/user/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 注册请求参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/user.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 返回token
          schema:
            type: string
      summary: 用户注册
      tags:
      - 用户相关接口
...
swagger: "2.0"